requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
werkzeug==3.0.1
